refactor() Moving files to domains if needed.
